feat(bot): Log bot into Steam account
Log the bot into its account on the Steam network. Support for
two-factor authentication has not been tested, and the bot configuration
needs to manually be created in "config/config.json".

Example of what the wrapper initially expects from the config file:

```
{
  "steam_id": "<steam_id>",
  "login": {
    "account_name": "<account_name>",
    "password": "<password>"
  }
}
```

After the first attempt returns an `EResult` response of `63`
(`AccountLogonDenied`), an authentication code (sent to the email
registered to the bot's account) is required:

```
{
  "steam_id": "<steam_id>",
  "login" {
    "account_name": "<account_name>",
    "password": "<password>",
    "auth_code": "<auth_code>"
  }
}
```