feat(client): Prompt user for Steam GUard code
The user is now prompted for the Steam Guard code and the bot no longer
requires a second, manual login attempt in order to provide it.

This means that `login.auth_code` has been removed as a required field
from the bot's config file. The bot's Steam ID also no longer needs to
be explicitly set and the `steam_id` option has also been removed.