feat(cmd): Add `dice` command (#18)
Add a command to simulate a dice roll. The command accepts a string
parameter of the form "#d#" where the first # is the number of dice to
roll and the second # is the max value of each die. The default roll is
`1d6` for a single, standard d6 die.

This includes a fix involving the parser short-circuit function; when
the parser passes an error, the arguments object is null and therefore
the context does not exist.