feat(bot): Allow custom config file
Require passing the path to the bot's config file. This allows for a set
of configuration settings specific to the bot.

The `confit` dependency has been replaced with `nconf`, as the former
doesn't seem to support custom config files. `confit` oddly does not
seem to support saving new values to the original file, either.